\name{AREsolver-package}
\alias{AREsolver-package}
\alias{AREsolver}
\docType{package}
\title{
\packageTitle{AREsolver}
}
\description{
\packageDescription{AREsolver}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AREsolver}
\packageIndices{AREsolver}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AREsolver}

Maintainer: \packageMaintainer{AREsolver}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
