% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiccatiSolver.R
\name{solvecare}
\alias{solvecare}
\title{Solving continuous time algebraic Riccati equation (CARE)}
\usage{
solvecare(A, B, Q, R)
}
\arguments{
\item{A}{matrix A, should be square}

\item{B}{matrix B, should be square}

\item{Q}{matrix Q, should be invertable}

\item{R}{matrix R, should be square}
}
\value{
matrix, the solution of the CARE, if algorithm failed, return NA
}
\description{
The continuous time algebraic Riccati equation (CARE) has form:
A'*X + P*X - XB(R^-1) B'X+Q = 0, where X is an unknown symmetric matrix.
This function uses eigenvector based method to return the stable solution
}
\examples{

my.X <- solvecare(diag(3),diag(3),diag(3),diag(3))
my.X + my.X-my.X \%*\% my.X


}
