% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiccatiSolver.R
\name{solvedare}
\alias{solvedare}
\title{Solving discrete time algebraic Riccati equation (DARE)}
\usage{
solvedare(A, B, Q, R)
}
\arguments{
\item{A}{matrix A, should be square}

\item{B}{matrix B, should be square}

\item{Q}{matrix Q, should be square}

\item{R}{matrix R, should be square}
}
\value{
matrix, the solution of the DARE, if algorithm failed, return NA
}
\description{
The discrete time algebraic Riccati equation (DARE) has form:
X=A'XA-(A'XB)(R+B'XB)^-1 (B'XA)+Q, where X is an unknown symmetric matrix.
This function uses eigenvector based method to return the stable solution
}
\examples{

my.X <- solvedare(diag(3),diag(3),diag(3),diag(3))

}
